requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6

# File processing
PyPDF2==3.0.1

# LangChain and AI dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-huggingface==0.0.1
sentence-transformers==2.2.2

# Vector store
faiss-cpu==1.7.4

# HuggingFace
huggingface-hub==0.19.4
transformers==4.36.2

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0